Move notification preferences logic to notification service
- Removed `create_or_update_notification_preferences` and `get_account_notification_preferences` from `AccountService`.
- Added equivalent methods to `NotificationService` using internal reader/writer.
- Clarifies service boundaries by centralizing notification handling.
