+++ src/apps/backend/modules/account/account_service.py
@@ -1,5 +1,3 @@
-from modules.notification.internals.account_notification_preferences_reader import AccountNotificationPreferenceReader
-from modules.notification.internals.account_notification_preferences_writer import AccountNotificationPreferenceWriter
 from modules.account.internal.account_reader import AccountReader
 from modules.account.internal.account_writer import AccountWriter
 from modules.account.types import (
@@ -21,12 +19,6 @@ class AccountService:
     @staticmethod
     def create_account_by_username_and_password(*, params: CreateAccountByUsernameAndPasswordParams) -> Account:
         return AccountWriter.create_account_by_username_and_password(params=params)
-
-    @staticmethod
-    def create_or_update_notification_preferences(
-        *, account_id: str, preferences: NotificationPreferencesParams
-    ) -> NotificationPreferencesParams:
-        return AccountNotificationPreferenceWriter.create_or_update_notification_preferences(account_id, preferences)
 
     @staticmethod
     def get_account_by_phone_number(*, phone_number: PhoneNumber) -> Account:
@@ -71,10 +63,6 @@ class AccountService:
     @staticmethod
     def get_account_by_username_and_password(*, params: AccountSearchParams) -> Account:
         return AccountReader.get_account_by_username_and_password(params=params)
-
-    @staticmethod
-    def get_account_notification_preferences(account_id: str) -> NotificationPreferencesParams:
-        return AccountNotificationPreferenceReader.get_notification_preferences_by_account_id(account_id)
 
     @staticmethod
     def update_account_profile(*, account_id: str, params: UpdateAccountProfileParams) -> Account:
